Rename `package mod` to `package pkg`
