+++ vpclogs/pkg/store_header.go
@@ -15,7 +15,7 @@
  *
  */
 
-package mod
+package pkg
 
 import (
 	"time"
